requirements.txt
langchain
langchain-classic
langchain-core
langchain-groq
langchain-mcp-adapters
langgraph
streamlit
wikipedia
ddgs
arxiv
fastapi
uvicorn
slowapi
python-dotenv
gunicorn
aiohttp
tavily-python
pydantic
pydantic-settings